refactor(vscode): extract server launch plumbing
