+++ editors/vscode/src/extension.ts
@@ -11,7 +11,6 @@ import {
   type ServerOptions,
 } from 'vscode-languageclient/node';
 
-import { getBundledServerPath, getPlatformFolder } from './platform';
 import { registerDiagnosticActions } from './diagnosticActions';
 import { profileDiagnosticsCommand, registerProfilingCommand } from './profiling';
 import { serverInitializationOptions } from './initializationOptions';
@@ -32,6 +31,11 @@ import {
   VideStatusController,
 } from './videStatus';
 import type { ServerStatus } from './status';
+import {
+  createServerEnv,
+  readConfiguration,
+  resolveServerLaunch,
+} from './node/serverLaunch';
 
 let client: LanguageClient | undefined;
 let outputChannel: vscode.OutputChannel | undefined;
@@ -289,107 +293,6 @@ function registerProjectStatusNotifications(languageClient: LanguageClient): voi
   });
 }
 
-interface ServerConfiguration {
-  command: string | undefined;
-  args: string[];
-  additionalArgs: string[];
-  cwd: string | undefined;
-  trace: 'off' | 'messages' | 'verbose';
-}
-
-interface ServerLaunch {
-  command: string;
-  args: string[];
-  additionalArgs: string[];
-  cwd: string;
-}
-
-function asStringArray(value: unknown): string[] | undefined {
-  return Array.isArray(value) && value.every((item) => typeof item === 'string')
-    ? value
-    : undefined;
-}
-
-function getServerPath(context: vscode.ExtensionContext): string | undefined {
-  const platform = process.platform;
-  const arch = process.arch;
-  const platformFolder = getPlatformFolder(platform, arch);
-  if (!platformFolder) {
-    log(
-      `[ERROR] Unsupported platform-architecture combination: ${platform}-${arch}`,
-    );
-    return undefined;
-  }
-
-  const bundledPath = getBundledServerPath(context.extensionPath, platform, arch);
-  if (!bundledPath) {
-    log(`[ERROR] Unsupported platform-architecture combination: ${platformFolder}`);
-    return undefined;
-  }
-
-  log(`[INFO] Looking for bundled server at: ${bundledPath}`);
-
-  if (fs.existsSync(bundledPath)) {
-    if (platform !== 'win32') {
-      try {
-        fs.accessSync(bundledPath, fs.constants.X_OK);
-        log('[INFO] Bundled server binary is executable');
-        return bundledPath;
-      } catch {
-        log(
-          '[WARN] Bundled server binary exists but is not executable, attempting to fix...',
-        );
-        try {
-          fs.chmodSync(bundledPath, 0o755);
-          log('[INFO] Made bundled server binary executable');
-          return bundledPath;
-        } catch (error) {
-          log(
-            `[ERROR] Failed to make bundled binary executable: ${(error as Error).message}`,
-          );
-        }
-      }
-    } else {
-      log('[INFO] Found bundled server binary');
-      return bundledPath;
-    }
-  } else {
-    log(`[INFO] Bundled server binary not found at: ${bundledPath}`);
-  }
-
-  return undefined;
-}
-
-function readConfiguration(): ServerConfiguration {
-  const config = vscode.workspace.getConfiguration('vide');
-  const command = config.get<string | null>('server.command');
-  const args = asStringArray(config.get<unknown>('server.args'));
-  const additionalArgs = asStringArray(config.get<unknown>('server.additionalArgs'));
-  const cwd = config.get<string | null>('server.cwd');
-  const trace = config.get<'off' | 'messages' | 'verbose'>('trace.server') ?? 'off';
-
-  if (!args || !additionalArgs) {
-    vscode.window.showErrorMessage(
-      vscode.l10n.t('vide server arguments settings must be arrays of strings.'),
-    );
-    return {
-      command: undefined,
-      args: [],
-      additionalArgs: [],
-      cwd: undefined,
-      trace,
-    };
-  }
-
-  return {
-    command: typeof command === 'string' && command.length > 0 ? command : undefined,
-    args,
-    additionalArgs,
-    cwd: typeof cwd === 'string' && cwd.length > 0 ? cwd : undefined,
-    trace,
-  };
-}
-
 function includeDeclarationInReferences(document: vscode.TextDocument): boolean {
   return (
     vscode.workspace
@@ -527,69 +430,6 @@ async function provideExpandedRenameEdits(
     token,
   );
   return await languageClient.protocol2CodeConverter.asWorkspaceEdit(edit as never, token);
-}
-
-function resolveWorkingDirectory(
-  context: vscode.ExtensionContext,
-  configuredCwd: string | undefined,
-): string {
-  if (configuredCwd) {
-    log(`[INFO] Using configured working directory: ${configuredCwd}`);
-    return configuredCwd;
-  }
-
-  const workspaceFolder = vscode.workspace.workspaceFolders?.[0];
-  if (workspaceFolder) {
-    const workspacePath = workspaceFolder.uri.fsPath;
-    log(`[INFO] Using workspace folder as working directory: ${workspacePath}`);
-    return workspacePath;
-  }
-
-  log(`[INFO] Using extension path as working directory: ${context.extensionPath}`);
-  return context.extensionPath;
-}
-
-function resolveServerLaunch(
-  context: vscode.ExtensionContext,
-  config: ServerConfiguration,
-): ServerLaunch {
-  const cwd = resolveWorkingDirectory(context, config.cwd);
-
-  let serverCommand = config.command;
-  if (!serverCommand) {
-    serverCommand = getServerPath(context);
-    if (!serverCommand) {
-      const message = vscode.l10n.t(
-        'Bundled Vide Language Server binary not found. Install the VSIX that matches your platform or configure "vide.server.command".',
-      );
-      log(`[ERROR] ${message}`);
-      throw new Error(message);
-    }
-  } else {
-    log(`[INFO] Using custom server command: ${serverCommand}`);
-  }
-
-  log(`[INFO] Server command: ${serverCommand}`);
-  log(`[INFO] Server args: ${JSON.stringify([...config.args, ...config.additionalArgs])}`);
-  log(`[INFO] Working directory: ${cwd}`);
-
-  return {
-    command: serverCommand,
-    args: config.args,
-    additionalArgs: config.additionalArgs,
-    cwd,
-  };
-}
-
-function createServerEnv(
-  logLevel: 'info' | 'debug' = 'info',
-  backtrace: '1' | 'full' = '1',
-): NodeJS.ProcessEnv {
-  return {
-    ...process.env,
-    RUST_BACKTRACE: backtrace,
-    RUST_LOG: logLevel,
-  };
 }
 
 type ProjectConfigTarget = {
@@ -762,7 +602,7 @@ async function createClient(context: vscode.ExtensionContext): Promise<LanguageC
   log('[INFO] Creating language client...');
 
   const config = readConfiguration();
-  const launch = resolveServerLaunch(context, config);
+  const launch = resolveServerLaunch(context, config, log);
   const serverArgs = [...launch.args, ...launch.additionalArgs];
 
   const commonEnv = {
@@ -866,7 +706,7 @@ async function restartClient(context: vscode.ExtensionContext): Promise<void> {
 async function showServerVersion(context: vscode.ExtensionContext): Promise<void> {
   try {
     const config = readConfiguration();
-    const launch = resolveServerLaunch(context, config);
+    const launch = resolveServerLaunch(context, config, log);
     const versionArgs = [...launch.args, '--version'];
     log(`[INFO] Checking server version: ${launch.command} ${versionArgs.join(' ')}`);
     const { stdout, stderr } = await execFileAsync(launch.command, versionArgs, {
@@ -1056,7 +896,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
   if (profileTraceEnabled) {
     context.subscriptions.push(
       registerProfilingCommand(context, {
-        resolveLaunch: () => resolveServerLaunch(context, readConfiguration()),
+        resolveLaunch: () => resolveServerLaunch(context, readConfiguration(), log),
         createEnv: createServerEnv,
       }),
     );
